+++ src/App.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { Person } from './components/Person/Person';
+import './App.scss';
+
+// eslint-disable-next-line no-unused-vars
+const misha = {
+  name: 'Misha',
+  age: 37,
+  sex: 'm',
+  isMarried: true,
+  partnerName: 'Natasha',
+};
+
+// eslint-disable-next-line no-unused-vars
+const olya = {
+  name: 'Olya',
+  sex: 'f',
+  isMarried: true,
+  partnerName: 'Maksym',
+};
+
+// eslint-disable-next-line no-unused-vars
+const alex = {
+  name: 'Alex',
+  age: 25,
+  sex: 'm',
+  isMarried: false,
+};
+
+const App = () => (
+  <div className="App">
+    <Person person={misha} />
+    <Person person={olya} />
+    <Person person={alex} />
+  </div>
+);
+
+export default App;
